+++ packages/kc_api/crates/kc-api-interface-adapter/src/from_trait/mst_use_item.rs
@@ -5,10 +5,10 @@ use std::collections::HashMap;
 
 impl From<Vec<kcapi_main::api_start2::get_data::ApiMstUseitem>> for InterfaceWrapper<MstUseItems> {
     fn from(use_items: Vec<kcapi_main::api_start2::get_data::ApiMstUseitem>) -> Self {
-        let mut item_map = HashMap::<i64, MstUseItem>::with_capacity(use_items.len());
+        let mut item_map = HashMap::<i32, MstUseItem>::with_capacity(use_items.len());
         for use_item in use_items {
             item_map.insert(
-                use_item.api_id,
+                use_item.api_id as i32,
                 InterfaceWrapper::<MstUseItem>::from(use_item).unwrap(),
             );
         }
@@ -21,7 +21,7 @@ impl From<Vec<kcapi_main::api_start2::get_data::ApiMstUseitem>> for InterfaceWra
 impl From<kcapi_main::api_start2::get_data::ApiMstUseitem> for InterfaceWrapper<MstUseItem> {
     fn from(use_item: kcapi_main::api_start2::get_data::ApiMstUseitem) -> Self {
         Self(MstUseItem {
-            id: use_item.api_id,
+            id: use_item.api_id as i32,
             name: use_item.api_name,
         })
     }
